fix: rename tasks app to periodic_engine
Directory, queue names, file paths, URL patterns, imports, fixtures, and dead Dockerfile stage all updated to match the periodic_engine name already declared in settings.py and apps.py.
