+++ backend/databus/asgi.py
@@ -12,7 +12,7 @@ import os
 from channels.routing import ProtocolTypeRouter, URLRouter
 from django.core.asgi import get_asgi_application
 
-from tasks.routing import websocket_urlpatterns
+from periodic_engine.routing import websocket_urlpatterns
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "databus.settings")
